+++ js/doc-snippets/src/multi-agent/simple.ts
@@ -15,7 +15,8 @@
  */
 
 import { gemini15Pro } from '@genkit-ai/googleai';
-import { genkit, z } from 'genkit';
+import { z } from 'genkit';
+import { genkit } from 'genkit/beta';
 
 const ai = genkit({});
