+++ js/genkit/src/genkit.ts
@@ -52,13 +52,11 @@ import {
   evaluate,
   generate,
   generateStream,
-  isExecutablePrompt,
   loadPromptFolder,
   prompt,
   rerank,
   retrieve,
 } from '@genkit-ai/ai';
-import { Chat, ChatOptions } from '@genkit-ai/ai/chat';
 import {
   EmbedderAction,
   EmbedderArgument,
@@ -99,13 +97,6 @@ import {
   defineSimpleRetriever,
   index,
 } from '@genkit-ai/ai/retriever';
-import {
-  Session,
-  SessionData,
-  SessionError,
-  SessionOptions,
-  getCurrentSession,
-} from '@genkit-ai/ai/session';
 import { ToolFn } from '@genkit-ai/ai/tool';
 import {
   Action,
@@ -122,7 +113,6 @@ import {
   z,
 } from '@genkit-ai/core';
 import { HasRegistry } from '@genkit-ai/core/registry';
-import { v4 as uuidv4 } from 'uuid';
 import { BaseEvalDataPointSchema } from './evaluator.js';
 import { logger } from './logging.js';
 import { GenkitPlugin, genkitPlugin } from './plugin.js';
@@ -350,6 +340,7 @@ export class Genkit implements HasRegistry {
 
     return executablePrompt;
   }
+
   /**
    * Defines and registers a prompt based on a function.
    *
@@ -747,114 +738,6 @@ export class Genkit implements HasRegistry {
     }
     return generateStream(this.registry, options);
   }
-
-  /**
-   * Create a chat session with the provided options.
-   *
-   * ```ts
-   * const chat = ai.chat({
-   *   system: 'talk like a pirate',
-   * })
-   * let response = await chat.send('tell me a joke')
-   * response = await chat.send('another one')
-   * ```
-   */
-  chat<I>(options?: ChatOptions<I>): Chat;
-
-  /**
-   * Create a chat session with the provided preabmle.
-   *
-   * ```ts
-   * const triageAgent = ai.definePrompt({
-   *   system: 'help the user triage a problem',
-   * })
-   * const chat = ai.chat(triageAgent)
-   * const { text } = await chat.send('my phone feels hot');
-   * ```
-   */
-  chat<I>(preamble: ExecutablePrompt<I>, options?: ChatOptions<I>): Chat;
-
-  /**
-   * Create a chat session with the provided options.
-   *
-   * ```ts
-   * const chat = ai.chat({
-   *   system: 'talk like a pirate',
-   * })
-   * let response = await chat.send('tell me a joke')
-   * response = await chat.send('another one')
-   * ```
-   */
-  chat<I>(
-    preambleOrOptions?: ChatOptions<I> | ExecutablePrompt<I>,
-    maybeOptions?: ChatOptions<I>
-  ): Chat {
-    let options: ChatOptions<I> | undefined;
-    let preamble: ExecutablePrompt<I> | undefined;
-    if (maybeOptions) {
-      options = maybeOptions;
-    }
-    if (preambleOrOptions) {
-      if (isExecutablePrompt(preambleOrOptions)) {
-        preamble = preambleOrOptions as ExecutablePrompt<I>;
-      } else {
-        options = preambleOrOptions as ChatOptions<I>;
-      }
-    }
-
-    const session = this.createSession();
-    if (preamble) {
-      return session.chat(preamble, options);
-    }
-    return session.chat(options);
-  }
-
-  /**
-   * Create a session for this environment.
-   */
-  createSession<S = any>(options?: SessionOptions<S>): Session<S> {
-    const sessionId = uuidv4();
-    const sessionData: SessionData = {
-      id: sessionId,
-      state: options?.initialState,
-    };
-    return new Session(this.registry, {
-      id: sessionId,
-      sessionData,
-      store: options?.store,
-    });
-  }
-
-  /**
-   * Loads a session from the store.
-   */
-  async loadSession(
-    sessionId: string,
-    options: SessionOptions
-  ): Promise<Session> {
-    if (!options.store) {
-      throw new Error('options.store is required');
-    }
-    const sessionData = await options.store.get(sessionId);
-
-    return new Session(this.registry, {
-      id: sessionId,
-      sessionData,
-      store: options.store,
-    });
-  }
-
-  /**
-   * Gets the current session from async local storage.
-   */
-  currentSession<S = any>(): Session<S> {
-    const currentSession = getCurrentSession(this.registry);
-    if (!currentSession) {
-      throw new SessionError('not running within a session');
-    }
-    return currentSession as Session;
-  }
-
   /**
    * A flow step that executes the provided function. Each run step is recorded separately in the trace.
    *
